refactor: move IntegerEncoding to own module
